+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0 - 7/9/23
+
+- Support for MySQL ~8.0
+- Refactor sqlite helpers to be general
+- Bugfix where disabling reconstruction resulted in an error
+- Allow counting disallowed columns
+- Bugfix api key not set when using openai completion method
+- Generalize tests to handle multiple sql dialects
+
 ## 0.1.4 - 7/5/23
 
 - Use an optional int instead of `math.inf` for an unrestricted SQL limit